feat: move commit orchestration from WTS to SnapEngine (Phase A)
commitSnap, commitMultiZoneSnap, uncommitSnap, applyBatchAssignments
moved from WindowTrackingService to SnapEngine. SnapIntent enum moved
to types.h. windowSnapStateChanged and windowFloatingClearedForSnap
signals moved to SnapEngine.

WTA D-Bus entry points (windowSnapped, windowSnappedMultiZone,
windowUnsnapped) and session-restore path rewired to call SnapEngine
directly. Signal connections wired in setEngines().

State writes still route through WTS (which propagates to SnapState)
during the transitional dual-store period — Phase D will collapse this.
