+++ src/CMakeLists.txt
@@ -175,6 +175,7 @@ set(plasmazones_core_SRCS
     # ScriptedAlgorithm machinery + builtins.qrc moved to libs/phosphor-tiles/.
     snap/SnapEngine.cpp
     snap/SnapEngine.h
+    snap/snapengine/commit.cpp
     snap/snapengine/float.cpp
     snap/snapengine/lifecycle.cpp
     snap/snapengine/navigation.cpp
