+++ src/dbus/windowtrackingadaptor/navigation.cpp
@@ -130,13 +130,13 @@ void WindowTrackingAdaptor::resnapToNewLayout()
 static bool processBatchEntries(WindowTrackingAdaptor* adaptor, const QVector<ZoneAssignmentEntry>& entries,
                                 const QString& action)
 {
-    WindowTrackingService* service = adaptor ? adaptor->service() : nullptr;
-    if (!service) {
+    SnapEngine* engine = adaptor ? adaptor->snapEngine() : nullptr;
+    if (!engine) {
         return false;
     }
 
-    WindowGeometryList geometries = service->applyBatchAssignments(
-        entries, WindowTrackingService::SnapIntent::UserInitiated, [adaptor]() -> QString {
+    WindowGeometryList geometries =
+        engine->applyBatchAssignments(entries, SnapIntent::UserInitiated, [adaptor]() -> QString {
             const QString cursor = adaptor->lastCursorScreenName();
             return cursor.isEmpty() ? adaptor->lastActiveScreenName() : cursor;
         });
